requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
mysqlclient==2.2.7
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
